backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
python-multipart==0.0.9
httpx==0.27.0
anthropic==0.34.0
google-genai==1.70.0
Pillow==10.4.0
stripe==10.12.0
supabase==2.7.4
psycopg2-binary==2.9.9
sqlalchemy==2.0.35
alembic==1.13.3
python-jose[cryptography]==3.3.0
python-dotenv==1.0.1
pydantic==2.9.0
pydantic-settings==2.5.2
elevenlabs==1.9.0
boto3==1.35.0
aiofiles==24.1.0
